+++ app/mailers/carpool_mailer.rb
@@ -1,10 +1,26 @@
 class CarpoolMailer < ApplicationMailer
-  def picked_up(user, carpool, driver_name)
+  def seat_offer(offer)
+    @user = offer.user
+    @carpool = offer.carpool
+    @driver_name = offer.ride.user.name
+    @carpool_url = carpool_url(@carpool)
+    mail to: @user.email, subject: "#{@driver_name} offered you a seat for #{@carpool.name}"
+  end
+
+  def offer_accepted(driver, carpool, rider_name)
+    @user = driver
+    @carpool = carpool
+    @rider_name = rider_name
+    @carpool_url = carpool_url(carpool)
+    mail to: driver.email, subject: "#{rider_name} accepted your seat offer for #{carpool.name}"
+  end
+
+  def removed_from_ride(user, carpool, driver_name)
     @user = user
     @carpool = carpool
     @driver_name = driver_name
     @carpool_url = carpool_url(carpool)
-    mail to: user.email, subject: "You have a seat in #{driver_name}'s ride for #{carpool.name}"
+    mail to: user.email, subject: "Your seat for #{carpool.name} was released"
   end
 
   def ride_canceled(user, carpool, driver_name)