+++ app/models/user.rb
@@ -2,6 +2,7 @@ class User < ApplicationRecord
   has_many :carpools, dependent: :restrict_with_error
   has_many :rides, dependent: :restrict_with_error
   has_many :ride_claims, dependent: :restrict_with_error
+  has_many :seat_offers, dependent: :destroy
 
   generates_token_for :magic_link, expires_in: 30.minutes